+++ packages/preact/src/index.ts
@@ -1,5 +1,6 @@
 export * from './adapters';
 export * from './components';
+export * from './exceptions';
 export { useForm, useFormStore } from './hooks';
 export * from './methods';
 export * from './transformation';