+++ packages/qwik/src/index.ts
@@ -1,6 +1,7 @@
 export * from './actions';
 export * from './adapters';
 export * from './components';
+export * from './exceptions';
 export * from './hooks';
 export * from './methods';
 export * from './transformation';